Django-Middleware-0x03/requirements.txt
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt>=4.0.1
blinker>=1.9
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
colorama==0.4.6
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
distro==1.9.0
Django==5.2.4
django-filter>=24.2
djangorestframework-simplejwt==5.3.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.2
jsonpatch>=1.33
jsonpointer==2.0
jsonschema>=4.19.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient>=2.2.0
netifaces==0.11.0
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.45.3
Pygments==2.17.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
Twisted==24.3.0
typing-extensions>=4.12.2,<5
urllib3==2.0.7
wadllib==1.3.6
wheel==0.42.0
zope.interface==6.1
python-dotenv==1.0.0